song author + song title on activity play + some bugs fixed + music queue BUG: connection needs to be not closed in addition to play music on queue correctly from p2p connection
